# Speed up huff0 table decode