--- a/huff0/decompress_test.go
+++ b/huff0/decompress_test.go
@@ -478,3 +478,41 @@ func BenchmarkDecompress4XNoTable(b *testing.B) {
 		})
 	}
 }
+
+func BenchmarkDecompress4XTable(b *testing.B) {
+	for _, tt := range testfiles {
+		test := tt
+		if test.err4X != nil {
+			continue
+		}
+		b.Run(test.name, func(b *testing.B) {
+			var s = &Scratch{}
+			s.Reuse = ReusePolicyNone
+			buf0, err := test.fn()
+			if err != nil {
+				b.Fatal(err)
+			}
+			if len(buf0) > BlockSizeMax {
+				buf0 = buf0[:BlockSizeMax]
+			}
+			compressed, _, err := Compress4X(buf0, s)
+			if err != test.err1X {
+				b.Fatal("unexpected error:", err)
+			}
+			s.Out = nil
+			b.ResetTimer()
+			b.ReportAllocs()
+			b.SetBytes(int64(len(buf0)))
+			for i := 0; i < b.N; i++ {
+				s, remain, err := ReadTable(compressed, s)
+				if err != nil {
+					b.Fatal(err)
+				}
+				_, err = s.Decompress4X(remain, len(buf0))
+				if err != nil {
+					b.Fatal(err)
+				}
+			}
+		})
+	}
+}