Big speedup on small blocks.

```
λ benchstat old.txt new2.txt
name                                   old time/op    new time/op    delta
Decompress4XTable/digits-12               243µs ± 0%     223µs ± 0%   -8.03%          (p=0.008 n=5+5)
Decompress4XTable/gettysburg-12          5.81µs ± 0%    4.84µs ± 0%  -16.70%          (p=0.008 n=5+5)
Decompress4XTable/twain-12                735µs ± 1%     687µs ± 1%   -6.53%          (p=0.008 n=5+5)
Decompress4XTable/low-ent.10k-12         73.1µs ± 1%    67.2µs ± 0%   -8.15%          (p=0.008 n=5+5)
Decompress4XTable/superlow-ent-10k-12    20.4µs ± 0%    18.7µs ± 1%   -8.04%          (p=0.008 n=5+5)
Decompress4XTable/case1-12               2.51µs ± 1%    2.15µs ± 1%  -14.44%          (p=0.008 n=5+5)
Decompress4XTable/case2-12               2.47µs ± 0%    2.09µs ± 1%  -15.12%          (p=0.008 n=5+5)
Decompress4XTable/case3-12               2.50µs ± 0%    2.11µs ± 0%  -15.53%          (p=0.008 n=5+5)
Decompress4XTable/pngdata.001-12         98.9µs ± 1%    91.8µs ± 0%   -7.13%          (p=0.008 n=5+5)
Decompress4XTable/normcount2-12          1.60µs ± 0%    1.49µs ± 0%   -6.84%          (p=0.008 n=5+5)

name                                   old speed      new speed      delta
Decompress4XTable/digits-12             412MB/s ± 0%   448MB/s ± 0%   +8.72%          (p=0.008 n=5+5)
Decompress4XTable/gettysburg-12         266MB/s ± 0%   320MB/s ± 0%  +20.04%          (p=0.008 n=5+5)
Decompress4XTable/twain-12              356MB/s ± 1%   381MB/s ± 1%   +6.99%          (p=0.008 n=5+5)
Decompress4XTable/low-ent.10k-12        547MB/s ± 1%   596MB/s ± 0%   +8.87%          (p=0.008 n=5+5)
Decompress4XTable/superlow-ent-10k-12   516MB/s ± 0%   561MB/s ± 1%   +8.74%          (p=0.008 n=5+5)
Decompress4XTable/case1-12             21.9MB/s ± 1%  25.6MB/s ± 1%  +16.89%          (p=0.008 n=5+5)
Decompress4XTable/case2-12             18.2MB/s ± 0%  21.5MB/s ± 1%  +17.81%          (p=0.008 n=5+5)
Decompress4XTable/case3-12             19.2MB/s ± 0%  22.7MB/s ± 0%  +18.39%          (p=0.008 n=5+5)
Decompress4XTable/pngdata.001-12        518MB/s ± 1%   558MB/s ± 0%   +7.68%          (p=0.008 n=5+5)
Decompress4XTable/normcount2-12        54.3MB/s ± 0%  58.3MB/s ± 0%   +7.32%          (p=0.008 n=5+5)
```